Update composer.json
This change adds a couple of development dependencies, sets the minimum
allowed version of PHP to 8.0, and adds a set of scripts to simplify
working with the project. The intent is to ensure the best quality code
in the project (though there's precious little), ensure it uses a
standard style, and is easy to verify.